# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Security AI Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "Api List Hosts",
        "description": "List enrolled hosts.\n\nBy default ``decommissioned`` hosts are hidden -- clicking 闁告帞濞€濞?makes\nthe host disappear from the table. The admin-only 鐎规瓕寮撶粭鍛棯婢剁鐦滈柡?view\npasses ``include_decommissioned=true`` to see the full roster.",
        "operationId": "api_list_hosts_api_v1_agents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "group",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group"
            }
          },
          {
            "name": "include_decommissioned",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "True=show soft-deleted hosts too (used by the admin 缂佺媴绱曢幃濠勬喆閸℃绂?; False=hide them so deletion actually removes the row from the list",
              "default": false,
              "title": "Include Decommissioned"
            },
            "description": "True=show soft-deleted hosts too (used by the admin 缂佺媴绱曢幃濠勬喆閸℃绂?; False=hide them so deletion actually removes the row from the list"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Api List Hosts Api V1 Agents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent_id}/actions/{action_name}": {
      "post": {
        "tags": [
          "response"
        ],
        "summary": "Dispatch Action",
        "description": "Dispatch one response action to ``agent_id``.\n\nBody shape:\n    { \"params\": { ... action-specific ... }, \"reason\": \"optional note\" }\n\nThe ``reason`` field is logged but not enforced -- operators sometimes\nwant a fast button without writing a justification.",
        "operationId": "dispatch_action_api_v1_agents__agent_id__actions__action_name__post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Agent Id"
            }
          },
          {
            "name": "action_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Action Name"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "title": "Body"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionDispatchResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/agents/actions/{action_id}": {
      "get": {
        "tags": [
          "response"
        ],
        "summary": "Get Action Status",
        "description": "Read the current status of a dispatched action.\n\nReturns ``status=\"unknown\"`` if no record is found (TTL expired, never\nexisted, or wrong id). Callers should treat ``unknown`` as\n\"investigate manually\" rather than retry.",
        "operationId": "get_action_status_api_v1_agents_actions__action_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "action_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 64,
              "title": "Action Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/agents/{agent_id}/monitor": {
      "get": {
        "tags": [
          "monitor"
        ],
        "summary": "List Monitor Events",
        "description": "Return the most recent N monitor snapshots for ``agent_id``.",
        "operationId": "list_monitor_events_api_v1_agents__agent_id__monitor_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "agent_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 128,
              "title": "Agent Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "description": "Max snapshots to return (most recent first)",
              "default": 20,
              "title": "Limit"
            },
            "description": "Max snapshots to return (most recent first)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Monitor Events Api V1 Agents  Agent Id  Monitor Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActionDispatchResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "action": {
            "type": "string",
            "title": "Action"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "dispatched | succeeded | failed | unknown",
            "default": "dispatched"
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "action",
          "agent_id"
        ],
        "title": "ActionDispatchResponse"
      },
      "ActionStatusResponse": {
        "properties": {
          "action_id": {
            "type": "string",
            "title": "Action Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "detail": {
            "type": "string",
            "title": "Detail",
            "default": ""
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id",
            "default": ""
          },
          "received_at": {
            "type": "string",
            "title": "Received At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "action_id",
          "status"
        ],
        "title": "ActionStatusResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}